Move simulatedimage.py -> artificial.py, remove pims
