+++ circletracking/__init__.py
@@ -3,8 +3,7 @@ from .refine import refine_ellipse, refine_ellipsoid_fast, refine_ellipsoid
 from .locate import locate_ellipse, locate_ellipsoid, locate_ellipsoid_fast
 from .algebraic import (ellipse_grid, ellipsoid_grid,
                         fit_ellipse, fit_ellipsoid)
-from .artificial import draw_ellipsoid, draw_ellipse
-from .simulatedimage import SimulatedImage
+from .artificial import draw_ellipsoid, draw_ellipse, SimulatedImage
 
 from ._version import get_versions
 __version__ = get_versions()['version']
